requirements.txt
opencv-python~=4.11.0.86
pillow~=11.1.0
streamlit~=1.42.0
pandas~=2.2.3